+++ HISTORY.rst
@@ -7,7 +7,9 @@ Pending release
 ---------------
 
 * New release notes go here
-* Added argument to pass custom ``TestDetails`` to ``record()``
+* Exposed the automatic naming logic used in ``record()`` in two new functions
+  ``get_perf_path()`` and ``get_record_name()``, in order to ease creation of
+  test records from calls outside of tests.
 
 2.0.1 (2017-03-02)
 ------------------